Refactor CustomTable parsing into separate partial and add test
Add a test for referencing custom tables and make the gigantic
Compiler.cs a bit smaller by extracting all the CustomTable handling
to separate partial.
